generateGoal/Floor/Robot additional arguments for customisation
Changed name of generateTarget to generateGoal for clarity.

Added additional arguments for generateGoal() that allows you to set a custom target position and the size of the target.
Added additional argument for generateFloor so that the size (scaling) can be set manually.
Added additional arguments for generateRobot so a custom pose can be used when initialising