Reward function revision
Errors fixed in how reward is calculated and some functionality changed based on change in spec.
Collision now trumps everything else if violated
time penalty is applied as a small penalty every timestep
fixes in logic of rotational and positional reward calculations

Constants added outside of reward method instead of within it.

Tolerances and vertical distance are set in the constructor, could

Some renaming of variables for clarity.